Moved rdump_factory to io/rdump.py and updated usages.
